+++ example/invaders/game.dart
@@ -5,6 +5,7 @@ import 'package:boarding/pieces.dart';
 import 'package:boarding/boarding.dart';
 import 'package:boarding/util.dart';
 
+part 'model/pieces.dart';
 part 'view/board.dart';
 
 main() {