+++ example/invaders/view/board.dart
@@ -4,50 +4,14 @@ class Board extends Surface {
   MovablePieces clouds, creatures;
   MovablePiece laser, spaceship;
   
-  Board(CanvasElement canvas) : super(canvas) {
-    clouds = new MovablePieces(5);
-    clouds.randomInit();
-    clouds.forEach((MovablePiece cloud) {
-      cloud.width = 70;
-      cloud.height = 70;
-      cloud.shape = PieceShape.IMG;
-      cloud.imageId = 'cloud';
-    });
-    
-    creatures = new MovablePieces(7);
-    creatures.randomInit();
-    creatures.forEach((MovablePiece creature) {
-      creature.width = 80;
-      creature.height = 80;
-      if (creature.dy < 2) {
-        creature.dy = 2;
-      }
-      if (creature.dx >= creature.dy) {
-        creature.dx = creature.dy - 1;
-      } 
-      creature.y = -creature.y;
-      var ri = randomInt(7);
-      creature.dx = randomSign(ri) * creature.dx; 
-      creature.shape = PieceShape.IMG;
-      creature.imageId = 'creature';
-    });
-    
-    laser = new MovablePiece();
-    laser.randomInit();
-    laser.width = 4;
-    laser.height = 50;
-    laser.color = 'gray';
-    laser.borderColor = 'red';
-    laser.speed.dy = 6;
-    laser.shape = PieceShape.RECT;
-    laser.isVisible = false;
-    
-    spaceship = new MovablePiece();
-    spaceship.randomInit();
-    spaceship.width = 70;
-    spaceship.height = 70;
-    spaceship.shape = PieceShape.IMG;
-    spaceship.imageId = 'spaceship';
+  AudioElement hitSound;
+  
+  Board(CanvasElement canvas) : super(canvas) {  
+    clouds = new Clouds(5);
+    creatures = new Creatures(7);
+    spaceship = new Spaceship();
+    laser = new Laser();
+    hitSound = document.querySelector('#${laser.audioId}');   
     
     canvas.onMouseMove.listen((MouseEvent e) {
       spaceship.x = e.offset.x - spaceship.width  / 2;
@@ -76,11 +40,11 @@ class Board extends Surface {
   draw() {
     clear();
     if (creatures.any((Piece p) => p.isVisible)) {
-      clouds.forEach((MovablePiece cloud) {
+      clouds.forEach((Cloud cloud) {
         cloud.move(Direction.DOWN);
         drawPiece(cloud);
       });
-      creatures.forEach((MovablePiece creature) {
+      creatures.forEach((Creature creature) {
         if (creature.isVisible) {
           creature.move(Direction.DOWN);
           creature.x += creature.dx; 
@@ -92,7 +56,9 @@ class Board extends Surface {
           }
           if (!creature.isSelected && laser.isVisible && laser.hit(creature)) {
             creature.isSelected = true;
-            creature.imageId = 'explosion';
+            creature.imgId = 'explosion';
+            hitSound.load();
+            hitSound.play();
             laser.isVisible = false;
           }
           drawPiece(creature);